# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel perfume api",
        "description": "L5 Swagger OpenApi ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/admin/factor": {
            "get": {
                "summary": "گرفتن تمام کتگوری های موجود برای ادمین",
                "operationId": "f2b4aa45ddf83537deea8c56ca44d46a",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "header",
                        "description": "احراز هویت با توکن",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/AdminFactor"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "unAuthorized"
                    }
                }
            },
            "post": {
                "summary": "اضافه کردن فاکتور",
                "operationId": "f216fc6fab4a5f8ad2620246ead781ba",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "header",
                        "description": "احراز هویت با توکن",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string"
                                                },
                                                "volume": {
                                                    "type": "string"
                                                },
                                                "price": {
                                                    "type": "string"
                                                },
                                                "quantity": {
                                                    "type": "string"
                                                },
                                                "slug": {
                                                    "type": "string"
                                                },
                                                "warranty": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "gender": {
                                                    "type": "string"
                                                },
                                                "percent": {
                                                    "type": "string"
                                                },
                                                "amount": {
                                                    "type": "string"
                                                },
                                                "start_date": {
                                                    "type": "string"
                                                },
                                                "end_date": {
                                                    "type": "string"
                                                },
                                                "discount_card": {
                                                    "type": "string"
                                                },
                                                "discount_card_percent": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "products": [
                                        {
                                            "name": "256",
                                            "volume": "29",
                                            "quantity": "510",
                                            "description": "this is fuking description",
                                            "price": "1000000",
                                            "slug": "lahg-asdgkjah-agh-askgh-lkjfg",
                                            "gender": "sport",
                                            "percent": "22.12",
                                            "end_date": "2011-03-03 14:14"
                                        },
                                        {
                                            "name": "256",
                                            "volume": "29",
                                            "quantity": "510",
                                            "description": "this is fuking description",
                                            "price": "1000000",
                                            "slug": "lahg-asdgkjah-agh-askgh-lkjfg",
                                            "gender": "sport",
                                            "percent": "22.12",
                                            "end_date": "2011-03-03 14:14"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "unAuthorized"
                    }
                }
            }
        },
        "/api/admin/factor/{id}": {
            "get": {
                "summary": "گرفتن اطلاعات مربوط به یک فاکتور خاص برای ادمین",
                "operationId": "f52f9513dfae6847f3d2abbdb2a7ad2f",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "header",
                        "description": "احراز هویت با توکن",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ایدی فاکتور مربوط در یو ار ال",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/AdminFactor"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "unAuthorized"
                    }
                }
            },
            "delete": {
                "summary": "حذف فاکتور موجود و مشتقاتش (فقط سوپر ادمین و سازنده فاکتور توانایی پاک کردن آن را دارند)",
                "operationId": "37a19bf4d5a8439c53e6583f4fd3047d",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "header",
                        "description": "احراز هویت با توکن",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ایدی فاکتور در یو ار ال",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "unAuthorized"
                    }
                }
            }
        },
        "/api/admin/factor/personal/{userId}": {
            "get": {
                "summary": "گرفتن فاکتور های  مربوط به یک ادمین خاص",
                "operationId": "074513bf192d56363a3781d5ea28ede2",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "header",
                        "description": "احراز هویت با توکن",
                        "required": true
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ایدی کاربر مربوط در یو ار ال",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/AdminFactor"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "unAuthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdminFactor": {
                "title": "همه فاکتور ها برای ادمین",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "isActive": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "string"
                    },
                    "userName": {
                        "type": "string"
                    },
                    "userRole": {
                        "type": "string"
                    },
                    "perfumeBasedFactor": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "volume": {
                                    "type": "string"
                                },
                                "price": {
                                    "type": "string"
                                },
                                "stock": {
                                    "type": "string"
                                },
                                "sold": {
                                    "type": "string"
                                },
                                "isActive": {
                                    "type": "string"
                                },
                                "warranty": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}